// packages/core/src/lib/cartesian/types.ts
export type boxAlignment =
    | 'center'
    | 'top-left'
    | 'top'
    | 'top-right'
    | 'right'
    | 'bottom-right'
    | 'bottom'
    | 'bottom-left'
    | 'left'

export type boxDimensions = {
    x: number
    y: number
    width: number
    height: number
}

export type xyCoordinates = [number, number]